extracted layout component
